Refactored bg messaging
