+++ background/messageHandlers/settings/general.js
@@ -1,0 +1,14 @@
+class GeneralHandlers {
+
+    static register(){
+        addBgMessageHandler("setSelectedCurrency", this.setSelectedCurrency)
+    }
+
+    static setSelectedCurrency(request, sender, sendResponse){
+        selectedCurrency = request.currency
+        browser.storage.local.set({"selectedCurrency": request.currency})
+    }
+
+}
+
+GeneralHandlers.register()
